+++ packages/react/src/atoms/Button/index.tsx
@@ -8,7 +8,7 @@ import { IconProp } from '@fortawesome/fontawesome-svg-core'
 import { ButtonStyled } from './styles'
 
 export type IButtonProps = React.ButtonHTMLAttributes<HTMLButtonElement> & {
-  variant?: 'primary' | 'secondary' | 'minimal' | 'danger' | 'link' | 'orange'
+  variant?: 'primary' | 'secondary' | 'minimal' | 'danger' | 'link'
   size?: 'small' | 'medium' | 'large'
   full?: boolean
   disabled?: boolean