+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @conquerpage/react
 
+## 2.0.2
+
+### Patch Changes
+
+- removes orange variant of the button
+
 ## 2.0.1
 
 ### Patch Changes